pyproject.toml
[project]
name = "drl-assignment-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argparse>=1.4.0",
    "dm-control==1.0.28",
    "gymnasium==1.1.1",
    "imageio==2.37.0",
    "numpy==2.2.5",
    "requests>=2.32.3",
    "shimmy==2.0.0",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
]
